CourseServiceImpl: added check for entitynotfound in delete method CourseServiceImpl: added check for entitynotfound in delete & edit methods MyCourseServiceImplTest: changed getMyCourse() test for above changes

+++ src/main/java/com/lens/coursetracker/service/CourseServiceImpl.java
@@ -44,6 +44,9 @@ public class CourseServiceImpl implements CourseService {
     @Override
     public void deleteById(int id) {
         logger.info("deleteById() -> " + id);
+        if(!courseRepository.findById(id).isPresent()){
+            throw new EntityNotFoundException("Course not found for id: " + id);
+        }
         courseRepository.deleteById(id);
     }
